+++ CHANGELOG.md
@@ -1,5 +1,26 @@
 # Version History
 
+## 0.2.0.dev0
+
+* Developer release (not public to pip)
+* Add support for datetime parsing with microseconds
+* Add support for datetime parsing with microseconds and timezone offset
+* Add support for time parsing with microseconds
+* Add pytests for datetime and time with microseconds
+* Add CSVReader.schema_overrides attribute
+* Add CSVReader.pandas_astype attribute
+* Add CSVReader.to_pandas() method
+* Add CSVReader.to_polars() method
+* Add CSVReader.to_bytes() method
+* Add CSVWriter.from_pandas() method
+* Add CSVWriter.from_polars() method
+* Add CSVWriter.from_bytes() method
+* Delete CSVPackReader.schema_overrides attribute
+* Delete CSVPackReader.pandas_astype attribute
+* Improve CSVReader.tell() method
+* Improve CSVWriter.tell() method
+* Remove CSVPackMeta.pandas_astype property method
+
 ## 0.1.0.dev6
 
 * Developer release (not public to pip)